_allocateMoreMemory function is broken
Currently allocated a really big amount so that we don't ever need to call the function.
will need to get fixed later on.

TODO: This is broken. freeing the memory is not a good idea because
there might be pointers pointed to the old memory that we plan on deleting.