Address INVALID_OPERATION error when running with nvidia card
This is just a temporary fix. I'm not fully sure what's going on.

Using this as a work around to make it so that the variables doesn't get
optimized out. It was causing INVALID_OPERATIONS
not 100% sure if this is the issue.
complaining for "uniform mat4 view"
doing this fails glUniformMatrix4fv(viewLoc, 1, GL_FALSE, glm::value_ptr(g_camera->view));