Windows: Able to compile with cmake for windows
Added make.bat to simulate previous setup.

Still need to fix on how to build the "render" portion

Fix issue where OpenGL context was not properly set.
We were passing the pointer argument as a value and not as a reference,
this caused it so that the actual thing we want is not being saved out
of the scope of the function. Use double pointer to save it.