build.bat: fix -set env before checking arguments.
I was setting the environment variables needed to build the renderer in
the main build path. If I never ran the main build path in the current
command prompt and then ran the render build. I will get a linker
failure.

Fix: put the LIB_PATH above the check so that it always get set.