Rework the way entities are rendered.
Refactor - instead of drawing each rect world position based on a
uniform variable 'position', we now generate all of the vertices with
the predetermined position. We no longer translate the position of a
rectangle in the shader.

Consolidate the draw calls. Instead of doing 1 draw call per entity --
thousands -- we accumulated all of the vertices in a buffer and draw
them in total of 2 calls. One for the debug program shader and one for
the normal program shader.

Refactor - Added unions to make things easier to pass around.

Refactor - clean up texture handling.