Add first pass of a "particle system"
Proof of concept for particle systems.
I'm still trying to figure out a good way to do this, but this is a good
stepping stone to get the ball rolling.

Ideally, we don't have to render all entites that is not close to the
screen. TBD.

Did some cleanup and removed dead code.