Fix Vulkan vkbuffer and vkdevicememory memory leaks.
I was not destroying the buffers that I was creating per vertices
update -- same goes for the memory used and created with VkDeviceMemory.

For now, I'm going to limit the amount of VkBuffer and VkDeviceMemory I
can create per frame to 100. Also, clean it up at the end of the render.