Vulkan: Update framerate stats once per frame.
This is sort of working -- still need to get some parts more robust.
One thint o think about is the size of the vkimage and memory.
Since the text is dynamic, the width of the texture image changes once
per frame.

There will be cases where the width of the vkimage is not enough and the
image gets cut-off.

TODO: Fix the issue mentioned above.