Vulkan: Add a new descriptor set for player texture.
We need to create a new image -> descriptor set -> to be able to use
vkBindDescriptorSet. This will change the currently used texture with a
new texture for the draw calls that happens afterwards.

This is a bit hacky -- hardcoding too many things. I will need to fix it
to be more automated.

misc: cleaned up some code