Fix sprite color. Fix projected z-axis for vulkan.
I'm not sure if the "Fix" for the z-axis for vulkan is valid. However,
OpenGL and Vulkan are now showing the exact same pixel lenght and width
for each object.

I still need to figure out why the screen coordinate space objects are
rendering different scales for Vulkan vs OpenGL.

Moved the "correction"/conversion of OpenGL -> Vulkan into the CPU side.
Before, we did it on the shader. I'm still not sure which one is better
to do.