Vulkan: make text overlay same size as OpenGL
The stats on the top left corner used to be disproportionate between
OpenGL and Vulkan. Fix it by creating a hack for Vulkan to ensure that
we only draw the texture that's been filled and not the empty parts of
the texture.

This is done by taking the difference between the generated string width
vs the max image width from vkimage that I previously created.