Scene is now drawing in correct order for Vulkan and OpenGL.
The current setup is still pretty manual, but at least it's drawing in
the correct order now -- back to front.

Refactor the code to add some functions when there's commonality between
my Vulkan and OpenGL code.

I'm not sure if this is the right way of doing things, still trying to
find the best way to do it as I go.