Use merge sort to fix the order of the draw objects
I need to draw from back to front, so I'll be using merge sort to have
the highest render layer at the top of the RenderGroup and the lowest
render layer at the end of the RenderGroup.

This might use a lot of memory, so I will also probably implement
quicksort.

This commit will double the amount of cycles to complete a frame.
Not acceptable at all...