Make entity manager into a dynamic array
Re-enable parts of the allocation for more memory once again.
Added a note related to why we can't free the memory at the moment.

Another reason we can't use `free` is because the pointer we're trying
to free is not the original base pointer. In otherwords, we're trying to
free an offset of the base pointer and it's causing a crash.