Fix white bitmap problem
white bitmap was not functioning properly before. It was basically doing
a no-opt because the whitebitmap was using a different descriptor set.
This basically did nothing in terms of showing something on the screen.

Fixed the order of the UI being drawn.
It should be the whitebitmap background first before the stats rect.

Added new texture and descriptor set for white bitmap on the vulkan
side.