Vulkan: Enable alpha blending
Enable alpha blending by tweaking the color attachment blend state.
This enables us to see through textures.

Tweak the shader to behave more to what is intended. I will probably
need to change this again later on.
May be use push constants to tell whether this is a debug rect or not?