Add SceneNode "QuadTree" into Kleos
This will enable us to query an AABB and get a list of rects within the
bounds.

This should help lessen the checks for collission detection.
This will fasten the navigation for AIs/NPCs.

Bug:
Fixed a bug with the dimension calculation. It used to account the full
height and width and ended up doubling the value when calculating the
`GetMinMax` -- use half the value instead by multiplying by 0.5f.