Refactor: Turn Array list into a macro
Delete the old implementation and use a macro array list.

The macro enforces the type when it expacts. We will no longer need to
check the type and address pointers and double pointers differently.

The previous ArrayList itteration was not straight forward... hopefully
the MACRO version works better.

The idea came from this blog:
https://solarianprogrammer.com/2017/01/08/c99-c11-dynamic-array-mimics-cpp-vector-api-improvements/

There are other things that still need to be implemented for the macro,
but the current one completely replaces the older ArrayList verison.

Had to use tripple pointers because we had to pass the double pointer as
a reference. This is needed so that we can reallocate the memory of the
original array instead of the copy (the double pointer copy that gets
generated when we pass the double pointer through the function).