Add Screen Space to World Space functionality.
We can now use the mouse and point at a specific location on the world
coordinates.

This is done by doing an "UnProject", which basically does the reverse
order of the normal MVP to get an image to the screen.

We will go from Screen Space -> Normalized Device Space ->
Clip Space -> View Space -> World Space
Take the inverse of the projection matrix to go back to view space.
Then take the inverse of the view space to go back to world space.

The value we get from the steps above will give us a ray.
This ray will need to be used to get an interection onto a plane to get
an actual point in the world space.

Normally we do:
Model -> World Space -> View Space -> Clip Space -> Normalized Device
Space -> Screen Space

Model being the default position of an entity in the local space.

- Added a bunch of visual debugging to sort of understand what's going
on.