Use generated bitmap font to show strings into view.
Each character will be a rect displayed on the screen.
The width and height of each character is hardcoded since we are not
using the glyph information.

We are rendering a string of text and then picking it apart to create a
"bitmap font". By default, the rendered strings are already glphyed, so
they are bunched up together.

Add "editmode".
 - When pressing `e`, this will trigger an "editmode" bar on the bottom.