Bug: Fix replay previous command.
There's a scenario where the command previous command is longer than the
current command. So when it's time to copy the current command, there
were lingering characters from the previous command.

Added a memset to clear out the whole buffer.

Previously, when you press the up key, the "cursor" will be at the
front.
Use string length to get to the end.