Bug: Fix intermittent issue with the GET_HASH and ADD_HASH
I was evaluating the address of the string as the key. This meant that
it will work with in the local scope. Once we are no longer using the
strings within the local scope, the KeyToHashIndex function will
eventually give a different value.

Trying to simplifying it further by getting rid of the (void *). Will
think about it further, so that we can have multiple
comparators/converters.