Fix hash delete macro
The previous hash delete macro was incomplete and had a couple of bugs.
Previously, it deleted all the children node if we wanted to delete a
parent node.

This is now fixed.

Added many Enemy NPCs to for testing.
This caused the frames to drop significantly -- this will be addressed.