Add spacial hash and hash sets for rect
Add spacial hash to prep up for better collission detection.
Added hash sets for rect to make it easier to find the right rect.
We are using a binary tree for the hash set (rect *).
This should be a red black tree for better balancing if there are issues
with a regular binary tree.